pyproject.toml
[tool.poetry]
name = "sync-blocked-email-domains"
version = "0.1.0"
description = "Automation to sync disposable email domains blocklist to Mastodon"
authors = ["Five-Borough-Fedi-Project"]
readme = "README.md"
packages = [{include = "sync_blocked_email_domains"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"

[tool.poetry.scripts]
sync-domains = "sync_blocked_email_domains.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
